+++ blocklib/zeromq/req_source/req_source_cpu.cc
@@ -29,7 +29,7 @@ work_return_code_t req_source_cpu::work(std::vector<block_work_input_sptr>& work
 
     auto noutput_items = work_output[0]->n_items;
     bool first = true;
-    int done = 0;
+    size_t done = 0;
 
     /* Process as much as we can */
     while (1) {
